refactor: Game Controller 관련 패키지 이동
