+++ BE/baseball-be/src/main/java/kr/codesquad/baseball/utils/SampleDataFactory.java
@@ -1,9 +1,9 @@
 package kr.codesquad.baseball.utils;
 
-import kr.codesquad.baseball.game.GameDTO;
-import kr.codesquad.baseball.game.GameScoresDTO;
-import kr.codesquad.baseball.game.LineUpDTO;
-import kr.codesquad.baseball.game.PlayerStatus;
+import kr.codesquad.baseball.game.controller.GameDTO;
+import kr.codesquad.baseball.game.controller.GameScoresDTO;
+import kr.codesquad.baseball.game.controller.LineUpDTO;
+import kr.codesquad.baseball.game.controller.PlayerStatus;
 import kr.codesquad.baseball.inning.BaseState;
 import kr.codesquad.baseball.inning.HitterRecord;
 import kr.codesquad.baseball.inning.InningDTO;
